requirements.txt
numpy==1.18.4
tensorflow==2.2.0
cv2==4.2.0
pandas==1.0.3
Werkzeug==1.0.1
Flask==1.1.2
Pillow==7.0.0